+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+## [1.0.0] - 2025-08-02
+### Changed
+- The default `equalsMethod` for `Equalone()` is now set to `deepEquals`.
+- The default value for `ignoreType` in `shallowEquals` and `deepEquals` is now `true`.
+- Updated examples.
+- Updated README.md.
+- Formatted code.
 
 ## [0.1.0] - 2025-07-27
 ### Added